Matched more functions in gcnewpause

+++ include/core2/1EA9160.h
@@ -2,6 +2,6 @@
 #define __CORE2_1EA9160_H__
 #include "common.h"
 
-u32 func_800D035C(u32 a0);
+u32 func_800D035C(s16 a0);
 u32 func_800D1338(u32 a0);
 #endif
